# Speedup SequenceMask on GPU