@haojin2 2019-03-16
@eric-haibin-lin @szha for review.