@haojin2 2019-03-19
@eric-haibin-lin @szha ping for review.